init refactor for local database support
